app/auth/jwt.py
from datetime import datetime, timedelta
from jose import jwt

SECRET_KEY = "SECRET"
ALGORITHM = "HS256"

def create_token(data: dict):
    to_encode = data.copy()
    to_encode["exp"] = datetime.utcnow() + timedelta(minutes=30)
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
