requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
python-dotenv
scikit-learn
joblib
python-jose
pytest
pytest-cov
locust
httpx
celery